+++ src/models/action/index.ts
@@ -8,3 +8,4 @@ export * from './temperatureSensorChangeAction';
 export { ActuatorChangeAction } from './actuatorChangeAction';
 export { BatteryLevelChangeAction } from './batteryLevelChangeAction';
 export { PresenceGroupAnyMovementAction } from './presenceGroupAnyMovementAction';
+export { HandleChangeAction } from './handleChangeAction';