+++ CHANGELOG.md
@@ -5,6 +5,11 @@
   ## **WORK IN PROGRESS**
   * (theimo1221) Update packages
 -->
+
+## **WORK IN PROGRESS**
+
+* (theimo1221) Shelly TRV set Window open on handle change.
+
 ## 3.0.0-beta.8 (2024-12-07)
 
 * (theimo1221) Harden HTTP Clients against connection errors.